+++ src/App.tsx
@@ -4,10 +4,14 @@ import WebRoutes from './WebRoutes'
 import 'antd/dist/antd.min.css'
 import { RecoilRoot } from 'recoil'
 import 'moment/locale/th'
+import { QueryClient, QueryClientProvider } from 'react-query'
+const queryClient = new QueryClient()
 function App() {
   return (
     <RecoilRoot>
-      <WebRoutes />
+      <QueryClientProvider client={queryClient}>
+        <WebRoutes />
+      </QueryClientProvider>
     </RecoilRoot>
   )
 }