+++ src/components/button/BackButton.tsx
@@ -4,16 +4,17 @@ import { color } from '../../resource'
 import { LeftOutlined } from '@ant-design/icons'
 
 interface BackBackProps {
-  onClick?: () => void
+  onClick?: () => void,
+  borderColor? : string
   disableBtn?: boolean
 }
 
-export const BackButton: React.FC<BackBackProps> = ({ onClick, disableBtn }) => (
+export const BackButton: React.FC<BackBackProps> = ({ onClick, disableBtn, borderColor }) => (
   <Button
     style={{
-      borderColor: color.Success,
+      borderColor: !borderColor ? color.Success : borderColor,
       borderRadius: '5px',
-      color: color.Success,
+      color: !borderColor ? color.Success : borderColor,
     }}
     onClick={onClick}
     disabled={disableBtn}